+++ kernel/net/fastsocket/Makefile
@@ -2,3 +2,6 @@
 # Makefile for FASTSOCKET module.
 #
 obj-$(CONFIG_FASTSOCKET) += fastsocket.o
+
+fastsocket-y := fastsocket_api.o fastsocket_core.o
+
