+++ kernel/net/fastsocket/fastsocket.h
@@ -7,7 +7,11 @@
 #ifndef _LINUX_FASTSOCKET_H
 #define _LINUX_FASTSOCKET_H
 
+#include <linux/ratelimit.h>
+#include <linux/fs.h>
+#include <linux/net.h>
 #include <linux/ioctl.h>
+#include <net/sock.h>
 
 #define IOC_ID 0xf5
 
@@ -25,7 +29,7 @@
 #define INFO 0x03
 #define DEBUG 0x04
 
-static int inline fsocket_get_dbg_level(void)
+static int inline fastsocket_dbg_level(void)
 {
 	//Declare the global variable inside this function to hide this varaible
 	extern int enable_fastsocket_debug;
@@ -33,40 +37,28 @@ static int inline fsocket_get_dbg_level(void)
 	return enable_fastsocket_debug;
 }
 
-DEFINE_RATELIMIT_STATE(fastsocket_ratelimit_state, 5 * HZ, 10);
 
 static inline int fastsocket_limit(void)
 {
+	//Declare the global variable inside this function to hide this varaible
+	extern struct ratelimit_state fastsocket_ratelimit_state;
+	
 	return __ratelimit(&fastsocket_ratelimit_state);
 }
 
 
 #ifndef DPRINTK
 #define DPRINTK(level, msg, args...) ({\
-	if (level < fsocket_get_dbg_level()) \
-		printk(KERN_DEBUG "Fastsocket [CPU%d][PID-%d] %s:%d\t" msg, smp_processor_id(), current->pid, __FUNCTION__, __LINE__, ## args); \
+	if (level < fastsocket_dbg_level()) \
+		printk(KERN_DEBUG "Fastsocket [CPU%d][PID-%d][%s] %s:%d\t" msg, smp_processor_id(), current->pid, current->comm, __FUNCTION__, __LINE__, ## args); \
 	})
 #endif
 
 #define EPRINTK_LIMIT(level, msg, args...) ({\
-	if (fastsocket_limit() && level < fsocket_get_dbg_level()) \
-		printk(KERN_DEBUG "Fastsocket [CPU%d][PID-%d] %s:%d\t" msg, smp_processor_id(), current->pid, __FUNCTION__, __LINE__, ## args); \
+	if (fastsocket_limit() && level < fastsocket_dbg_level()) \
+		printk(KERN_DEBUG "Fastsocket [CPU%d][PID-%d][%s] %s:%d\t" msg, smp_processor_id(), current->pid, current->comm, __FUNCTION__, __LINE__, ## args); \
 	})
 
-struct fsocket_alloc {
-	struct socket socket;
-	struct inode vfs_inode;
-};
-
-static inline struct inode *SOCKET_INODE(struct socket *socket)
-{
-	return &container_of(socket, struct fsocket_alloc, socket)->vfs_inode;
-}
-
-static inline struct socket *INODE_SOCKET(struct inode *inode)
-{
-	return &container_of(inode, struct socket_alloc, vfs_inode)->socket;
-}
 
 struct fsocket_ioctl_arg {
 	u32 fd;
@@ -106,5 +98,39 @@ struct fsocket_ioctl_arg {
 		}epoll_op;
 	}op;
 };
+
+/*
+Fastsocket global variable
+*/
+extern struct vfsmount *fastsocket_mnt;
+/* Fastsocket feature switches */
+extern int enable_listen_spawn;
+extern int enable_receive_flow_deliver;
+extern int enable_fast_epoll;
+extern int enable_skb_pool;
+extern int enable_rps_framework;
+extern int enable_receive_cpu_selection;
+extern int enable_direct_tcp;
+/* Fastsocket spawn cpu */
+extern cpumask_t fastsocket_spawn_cpuset;
+extern int fastsocket_spawn_cpu;
+extern struct mutex fastsocket_spawn_mutex;
+
+/*
+Fastsocket core functions
+*/
+extern int fsocket_init(void);
+extern void fsocket_exit(void);
+
+extern int fsocket_socket(int flags);
+extern int fsocket_listen(struct file *file, int backlog);
+extern int fsocket_spawn(struct file *filp, int fd, int tcpu);
+extern int fsocket_accept(struct file *file , struct sockaddr __user *upeer_sockaddr,
+		int __user *upeer_addrlen, int flags);
+extern int fsocket_shutdown_listen(struct file *file, int how);
+extern int fsocket_close(unsigned int fd);
+
+extern struct inode *fsocket_alloc_inode(struct super_block *sb);
+extern void fsocket_destroy_inode(struct inode *inode);
 
 #endif
